+++ plugins/module_utils/oci_version.py
@@ -4,4 +4,4 @@
 # Apache License v2.0
 # See LICENSE.TXT for details.
 
-__version__ = "4.23.0"
+__version__ = "4.24.0"