+++ plugins/modules/oci_management_dashboard_management_saved_search.py
@@ -54,13 +54,14 @@ options:
         type: str
     provider_version:
         description:
-            - Version of the service that owns this saved search.
+            - The version of the metadata of the provider. This is useful for provider to version its features and metadata. Any newly created saved search (or
+              dashboard) should use providerVersion 3.0.0.
             - Required for create using I(state=present).
             - This parameter is updatable.
         type: str
     provider_name:
         description:
-            - Name of the service (for example, Logging Analytics) that owns the saved search.
+            - The user friendly name of the service (for example, Logging Analytics) that owns the saved search.
             - Required for create using I(state=present).
             - This parameter is updatable.
         type: str
@@ -106,13 +107,14 @@ options:
             - "FILTER_DONT_SHOW_IN_DASHBOARD"
     ui_config:
         description:
-            - JSON that contains user interface options.
+            - It defines the visualization type of the widget saved search, the UI options of that visualization type, the binding of data to the visualization.
             - Required for create using I(state=present).
             - This parameter is updatable.
         type: dict
     data_config:
         description:
-            - Array of JSON that contain data source options.
+            - It defines how data is fetched. A functional saved search needs a valid dataConfig. See examples on how it can be constructed for various data
+              sources.
             - Required for create using I(state=present).
             - This parameter is updatable.
         type: list
@@ -125,19 +127,19 @@ options:
         type: str
     metadata_version:
         description:
-            - Version of the metadata.
+            - The version of the metadata defined in the API. This is maintained and enforced by dashboard server. Currently it is 2.0.
             - Required for create using I(state=present).
             - This parameter is updatable.
         type: str
     widget_template:
         description:
-            - Reference to the HTML file of the widget.
+            - The UI template that the saved search uses to render itself.
             - Required for create using I(state=present).
             - This parameter is updatable.
         type: str
     widget_vm:
         description:
-            - Reference to the view model of the widget.
+            - The View Model that the saved search uses to render itself.
             - Required for create using I(state=present).
             - This parameter is updatable.
         type: str
@@ -147,6 +149,11 @@ options:
             - This parameter is updatable.
         type: list
         elements: dict
+    features_config:
+        description:
+            - Contains configuration for enabling features.
+            - This parameter is updatable.
+        type: dict
     drilldown_config:
         description:
             - Drill-down configuration to define the destination of a drill-down action.
@@ -207,6 +214,7 @@ EXAMPLES = """
     # optional
     id: "ocid1.resource.oc1..xxxxxxEXAMPLExxxxxx"
     parameters_config: [ "parameters_config_example" ]
+    features_config: null
     drilldown_config: [ "drilldown_config_example" ]
     freeform_tags: {'Department': 'Finance'}
     defined_tags: {'Operations': {'CostCenter': 'US'}}
@@ -233,6 +241,7 @@ EXAMPLES = """
     widget_template: widget_template_example
     widget_vm: widget_vm_example
     parameters_config: [ "parameters_config_example" ]
+    features_config: null
     drilldown_config: [ "drilldown_config_example" ]
     freeform_tags: {'Department': 'Finance'}
     defined_tags: {'Operations': {'CostCenter': 'US'}}
@@ -258,6 +267,7 @@ EXAMPLES = """
     widget_template: widget_template_example
     widget_vm: widget_vm_example
     parameters_config: [ "parameters_config_example" ]
+    features_config: null
     drilldown_config: [ "drilldown_config_example" ]
     freeform_tags: {'Department': 'Finance'}
     defined_tags: {'Operations': {'CostCenter': 'US'}}
@@ -347,25 +357,28 @@ management_saved_search:
             sample: SEARCH_SHOW_IN_DASHBOARD
         ui_config:
             description:
-                - JSON that contains user interface options.
+                - It defines the visualization type of the widget saved search, the UI options of that visualization type, the binding of data to the
+                  visualization.
             returned: on success
             type: dict
             sample: {}
         data_config:
             description:
-                - Array of JSON that contain data source options.
+                - It defines how data is fetched. A functional saved search needs a valid dataConfig. See examples on how it can be constructed for various data
+                  sources.
             returned: on success
             type: list
             sample: []
         created_by:
             description:
-                - User who created the saved search.
+                - The principle id of the user that created this saved search. This is automatically managed by the system. In OCI the value is ignored. In EM
+                  it can skipped or otherwise it is ignored in both create and update API and system automatically sets its value.
             returned: on success
             type: str
             sample: created_by_example
         updated_by:
             description:
-                - User who updated the saved search.
+                - The principle id of the user that updated this saved search.
             returned: on success
             type: str
             sample: updated_by_example
@@ -389,25 +402,25 @@ management_saved_search:
             sample: screen_image_example
         metadata_version:
             description:
-                - Version of the metadata.
+                - The version of the metadata defined in the API. This is maintained and enforced by dashboard server. Currently it is 2.0.
             returned: on success
             type: str
             sample: metadata_version_example
         widget_template:
             description:
-                - Reference to the HTML file of the widget.
+                - The UI template that the saved search uses to render itself.
             returned: on success
             type: str
             sample: widget_template_example
         widget_vm:
             description:
-                - Reference to the view model of the widget.
+                - The View Model that the saved search uses to render itself.
             returned: on success
             type: str
             sample: widget_vm_example
         lifecycle_state:
             description:
-                - State of dashboard.
+                - OCI lifecycle status. This is automatically managed by the system.
             returned: on success
             type: str
             sample: ACTIVE
@@ -417,6 +430,12 @@ management_saved_search:
             returned: on success
             type: list
             sample: []
+        features_config:
+            description:
+                - Contains configuration for enabling features.
+            returned: on success
+            type: dict
+            sample: {}
         drilldown_config:
             description:
                 - Drill-down configuration to define the destination of a drill-down action.
@@ -460,6 +479,7 @@ management_saved_search:
         "widget_vm": "widget_vm_example",
         "lifecycle_state": "ACTIVE",
         "parameters_config": [],
+        "features_config": {},
         "drilldown_config": [],
         "freeform_tags": {'Department': 'Finance'},
         "defined_tags": {'Operations': {'CostCenter': 'US'}}
@@ -661,6 +681,7 @@ def main():
             widget_template=dict(type="str"),
             widget_vm=dict(type="str"),
             parameters_config=dict(type="list", elements="dict"),
+            features_config=dict(type="dict"),
             drilldown_config=dict(type="list", elements="dict"),
             freeform_tags=dict(type="dict"),
             defined_tags=dict(type="dict"),